typeId to Long

+++ src/main/java/com/gmail/netcracker/application/config/RootConfig.java
@@ -290,7 +290,7 @@ public class RootConfig {
             event.setCreator(getLong(rs, "creator_id"));
             event.setDateStart(getString(rs, "start_date"));
             event.setDateEnd(getString(rs, "end_date"));
-            event.setTypeId(getInt(rs, "type_id"));
+            event.setTypeId(getLong(rs, "type_id"));
             event.setDraft(getBoolean(rs, "is_draft"));
             event.setFolder(getInt(rs, "folder_id"));
             event.setWidth(getDouble(rs, "latitude"));
